updating_modules_file_structure
