+++ main.go
@@ -1,12 +1,16 @@
 package main
 
 import (
-	database "cloudflaretinyurl"
 	"log"
 	"net/http"
 
-	"tinyurl-api/database"
-	"tinyurl-api/routes"
+	"cloudflaretinyurl/database"
+	"cloudflaretinyurl/rediscounter"
+	"cloudflaretinyurl/redislocks"
+	"cloudflaretinyurl/redispubsub"
+	"cloudflaretinyurl/redisqueue"
+	"cloudflaretinyurl/routes"
+	"cloudflaretinyurl/utils"
 )
 
 func main() {
@@ -14,6 +18,23 @@ func main() {
 	if err := database.InitDB(); err != nil {
 		log.Fatalf("Initialization Error: %v", err)
 	}
+
+	// Initialize Snowflake ID generator with machine ID 1
+	if err := utils.InitSnowflake(1); err != nil {
+		log.Fatalf("Failed to initialize Snowflake ID generator: %v", err)
+	}
+
+	// Initialize Redis-based services (Counters, Queues, Pub/Sub, Locks)
+	rediscounter.InitRedisCounter(database.RDB)
+	redisqueue.InitRedisQueue(database.RDB)
+	redispubsub.InitRedisPubSub(database.RDB)
+	redislocks.InitRedisLocks(database.RDB)
+
+	// Start processing expired clicks in a separate goroutine
+	go redisqueue.ProcessExpiredClicks()
+
+	// Start listening for Redis Pub/Sub events
+	go redispubsub.ListenForExpiredClicks()
 
 	// Set up API routes
 	r := routes.InitRoutes()
